requirements.txt
azure-ai-agentserver-agentframework
azure-ai-projects
azure-identity
openai
requests
beautifulsoup4
python-dotenv
